chore: :truck: consolidate many of the path functions into fewer files
