+++ src/seedcase_sprout/core/path_package_functions.py
@@ -1,16 +1,20 @@
 """This module contains functions to get the paths to data packages and their files.
 
 They are intended to be used in conjunction with other functions to read, write, and
-edit the contents and properties of packages.
+edit the contents and properties of packages. These functions in particular are used
+for the global location of packages, in environments with potentially multiple users
+and many data packages.
 """
 
+from os import getenv
 from pathlib import Path
+
+from platformdirs import user_data_path
 
 from seedcase_sprout.core.check_is_dir import check_is_dir
 from seedcase_sprout.core.check_is_file import check_is_file
 from seedcase_sprout.core.check_is_package_dir import check_is_package_dir
 from seedcase_sprout.core.create_dirs import create_dir
-from seedcase_sprout.core.path_sprout_global import path_sprout_global
 
 
 def path_package(package_id: int) -> Path:
@@ -151,3 +155,51 @@ def path_readme(package_id: int) -> Path:
     """
     path = path_package(package_id=package_id) / "README.md"
     return check_is_file(path)
+
+
+def path_sprout_global() -> Path:
+    """Gets Sprout's global path location.
+
+    If the `SPROUT_GLOBAL` environment variable isn't provided, this function
+    will return the default path to where data packages will be stored. The
+    default locations are dependent on the operating system.  This function also
+    creates the necessary directory if it doesn't exist.
+
+    Returns:
+        The path to Sprout's global directory.
+
+    Examples:
+        ```{python}
+        import tempfile
+        import os
+
+        import seedcase_sprout.core as sp
+
+        # Create a temporary directory for the example
+        with tempfile.TemporaryDirectory() as temp_dir:
+            os.environ["SPROUT_GLOBAL"] = temp_dir
+
+            # Get the path to Sprout's global directory
+            sp.path_sprout_global()
+        ```
+    """
+    return _get_sprout_global_envvar() or _create_sprout_global_path()
+
+
+def _create_sprout_global_path() -> Path:
+    """Creates the path to Sprout global location.
+
+    Returns:
+        The path with the Sprout global directory tied to the user.
+    """
+    return user_data_path("sprout")
+
+
+def _get_sprout_global_envvar() -> Path | None:
+    """Gets the global environment variable `SPROUT_GLOBAL` if it exists.
+
+    Returns:
+        The path containing `SPROUT_GLOBAL` if it is set, otherwise None.
+    """
+    sprout_global = getenv("SPROUT_GLOBAL")
+    return Path(sprout_global) if sprout_global else None
